policy-cartography/pyproject.toml
[project]
name = "policy-cartography"
version = "0.1.0"
description = "State-space policy-boundary cartography sandbox for tool-using LLM agents"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "requests>=2.32",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5",
  "mypy>=1.10",
]

[tool.ruff]
line-length = 100
